+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 更改日志 | CHANGE LOG
 
 - 修复复制当前文档大纲为有序列表时序号错误问题 | Fixed the issue that the number of the list was wrong when copying the current document outline as an ordered list.
+- 视频时间戳支持 Bilibili 与 YouTube | Support the timestamp of videos on Bilibili and YouTube.
 
 ## v0.5.4/2022-03-29